SCRUM-1493: use panache rest extension automating CRUD

+++ src/main/java/org/alliancegenome/mati/repository/SubdomainRepository.java
@@ -6,4 +6,4 @@ import org.alliancegenome.mati.entity.SubdomainEntity;
 import javax.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
-public class SubdomainRepository implements PanacheRepositoryBase<SubdomainEntity, Integer> { }
+public class SubdomainRepository implements PanacheRepositoryBase<SubdomainEntity, Long> { }
